+++ src/api.rs
@@ -16,13 +16,14 @@
 use crate::error::Error;
 use crate::error::Error::{ClientErrorResponse, NotFound, ServerErrorResponse};
 use crate::requests::{
-    Amqp091ShovelParams, Amqp10ShovelParams, EmptyPayload, FederationUpstreamParams, StreamParams,
-    FEDERATION_UPSTREAM_COMPONENT, SHOVEL_COMPONENT,
+    Amqp091ShovelParams, Amqp10ShovelParams, EmptyPayload, FederationUpstreamParams,
+    GlobalRuntimeParameterDefinition, StreamParams, FEDERATION_UPSTREAM_COMPONENT,
+    SHOVEL_COMPONENT,
 };
 use crate::responses::{
-    DeprecatedFeatureList, FeatureFlag, FeatureFlagList, FeatureFlagStability, FeatureFlagState,
-    FederationUpstream, GetMessage, OAuthConfiguration, Overview, SchemaDefinitionSyncStatus,
-    VirtualHostDefinitionSet, WarmStandbyReplicationStatus,
+    ClusterTags, DeprecatedFeatureList, FeatureFlag, FeatureFlagList, FeatureFlagStability,
+    FeatureFlagState, FederationUpstream, GetMessage, OAuthConfiguration, Overview,
+    SchemaDefinitionSyncStatus, VirtualHostDefinitionSet, WarmStandbyReplicationStatus,
 };
 use crate::{
     commons::{BindingDestinationType, SupportedProtocol, UserLimitTarget, VirtualHostLimitTarget},
@@ -1040,6 +1041,42 @@ where
         Ok(())
     }
 
+    pub async fn list_global_runtime_parameters(
+        &self,
+    ) -> Result<Vec<responses::GlobalRuntimeParameter>> {
+        let response = self.http_get("global-parameters", None, None).await?;
+        let response = response.json().await?;
+        Ok(response)
+    }
+
+    pub async fn get_global_runtime_parameter(
+        &self,
+        name: &str,
+    ) -> Result<responses::GlobalRuntimeParameter> {
+        let response = self
+            .http_get(path!("global-parameters", name), None, None)
+            .await?;
+        let response = response.json().await?;
+        Ok(response)
+    }
+
+    pub async fn upsert_global_runtime_parameter<'a>(
+        &self,
+        param: &'a GlobalRuntimeParameterDefinition<'a>,
+    ) -> Result<()> {
+        let _response = self
+            .http_put(path!("global-parameters", param.name), &param, None, None)
+            .await?;
+        Ok(())
+    }
+
+    pub async fn clear_global_runtime_parameter(&self, name: &str) -> Result<()> {
+        let _response = self
+            .http_delete(path!("global-parameters", name), None, None)
+            .await?;
+        Ok(())
+    }
+
     pub async fn set_user_limit(
         &self,
         username: &str,
@@ -1127,6 +1164,25 @@ where
     pub async fn set_cluster_name(&self, new_name: &str) -> Result<()> {
         let body = json!({"name": new_name});
         let _response = self.http_put("cluster-name", &body, None, None).await?;
+        Ok(())
+    }
+
+    pub async fn get_cluster_tags(&self) -> Result<responses::ClusterTags> {
+        let response = self.get_global_runtime_parameter("cluster_tags").await?;
+        Ok(ClusterTags::from(response.value))
+    }
+
+    pub async fn set_cluster_tags(&self, tags: Map<String, Value>) -> Result<()> {
+        let grp = GlobalRuntimeParameterDefinition {
+            name: "cluster_tags",
+            value: tags,
+        };
+        self.upsert_global_runtime_parameter(&grp).await?;
+        Ok(())
+    }
+
+    pub async fn clear_cluster_tags(&self) -> Result<()> {
+        self.clear_global_runtime_parameter("cluster_tags").await?;
         Ok(())
     }