+++ tests/async_runtime_parameter_tests.rs
@@ -11,13 +11,17 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-use rabbitmq_http_client::requests::{RuntimeParameterDefinition, RuntimeParameterValue};
+use rabbitmq_http_client::requests::{
+    GlobalRuntimeParameterDefinition, RuntimeParameterDefinition,
+};
 use rabbitmq_http_client::responses::RuntimeParameter;
 use rabbitmq_http_client::{api::Client, requests::VirtualHostParams};
 use serde_json::{json, Map, Value};
 
 mod test_helpers;
-use crate::test_helpers::{async_await_metric_emission, endpoint, PASSWORD, USERNAME};
+use crate::test_helpers::{
+    async_await_metric_emission, cluster_tags, endpoint, PASSWORD, USERNAME,
+};
 
 #[tokio::test]
 async fn test_async_upsert_runtime_parameter() {
@@ -47,6 +51,69 @@ async fn test_async_upsert_runtime_parameter() {
             .as_u64()
             .unwrap()
     );
+
+    let _ = rc
+        .clear_runtime_parameter(rpf.component, rpf.vhost, rpf.name)
+        .await;
+    let _ = rc.delete_vhost(vh_params.name, false).await;
+}
+
+#[tokio::test]
+async fn test_async_list_all_runtime_parameters() {
+    let endpoint = endpoint();
+    let rc = Client::new(&endpoint, USERNAME, PASSWORD);
+
+    let vh_params =
+        VirtualHostParams::named("rust/http/api/async/test_async_list_runtime_parameters");
+    let result1 = rc.create_vhost(&vh_params).await;
+    assert!(result1.is_ok());
+
+    let mut val = max_connections_limit(9988);
+    let rpf = example_runtime_parameter_definition(vh_params.name, &mut val);
+    let result2 = rc.upsert_runtime_parameter(&rpf).await;
+    assert!(result2.is_ok());
+
+    let result3 = rc.list_runtime_parameters().await;
+    dbg!(&result3);
+    assert!(result3.is_ok());
+    assert!(result3
+        .unwrap()
+        .iter()
+        .filter(|rp| rp.component == "vhost-limits" && rp.vhost == *vh_params.name)
+        .map(|rp| rp.value.get("max-connections").unwrap().as_u64().unwrap())
+        .any(|n| n == 9988));
+
+    let _ = rc
+        .clear_runtime_parameter(rpf.component, rpf.vhost, rpf.name)
+        .await;
+    let _ = rc.delete_vhost(vh_params.name, false).await;
+}
+
+#[tokio::test]
+async fn test_async_list_runtime_parameters_of_component_in_a_vhost() {
+    let endpoint = endpoint();
+    let rc = Client::new(&endpoint, USERNAME, PASSWORD);
+
+    let vh_params = VirtualHostParams::named(
+        "rust/http/api/async/test_async_list_runtime_parameters_of_component_in_a_vhost",
+    );
+    let result1 = rc.create_vhost(&vh_params).await;
+    assert!(result1.is_ok());
+
+    let mut val = max_connections_limit(9988);
+    let rpf = example_runtime_parameter_definition(vh_params.name, &mut val);
+    let result2 = rc.upsert_runtime_parameter(&rpf).await;
+    assert!(result2.is_ok());
+
+    let result3 = rc
+        .list_runtime_parameters_of_component_in("vhost-limits", vh_params.name)
+        .await;
+    assert!(result3.is_ok());
+    assert!(result3
+        .unwrap()
+        .iter()
+        .map(|rp| rp.value.get("max-connections").unwrap().as_u64().unwrap())
+        .any(|n| n == 9988));
 
     let _ = rc
         .clear_runtime_parameter(rpf.component, rpf.vhost, rpf.name)
@@ -88,6 +155,54 @@ async fn test_async_clear_runtime_parameter() {
 }
 
 #[tokio::test]
+async fn test_async_upsert_global_runtime_parameter() {
+    let endpoint = endpoint();
+    let rc = Client::new(&endpoint, USERNAME, PASSWORD);
+
+    let mut tags: Map<String, Value> = Map::new();
+    tags.insert("region".to_owned(), json!("ca-central-1"));
+    let grp = GlobalRuntimeParameterDefinition {
+        name: "cluster_tags",
+        value: cluster_tags(tags),
+    };
+
+    let result1 = rc.upsert_global_runtime_parameter(&grp).await;
+    assert!(result1.is_ok());
+
+    let result2 = rc.get_global_runtime_parameter("cluster_tags").await;
+    assert!(result2.is_ok());
+    assert_eq!(
+        result2.unwrap().value.0.get("cluster_tags").unwrap(),
+        &json!({"region": "ca-central-1"})
+    );
+
+    let result3 = rc.clear_global_runtime_parameter("cluster_tags").await;
+    assert!(result3.is_ok());
+}
+
+#[tokio::test]
+async fn test_async_list_global_runtime_parameters() {
+    let endpoint = endpoint();
+    let rc = Client::new(&endpoint, USERNAME, PASSWORD);
+
+    let mut tags: Map<String, Value> = Map::new();
+    tags.insert("region".to_owned(), json!("ca-central-1"));
+    let grp = GlobalRuntimeParameterDefinition {
+        name: "cluster_tags",
+        value: cluster_tags(tags),
+    };
+
+    let result1 = rc.upsert_global_runtime_parameter(&grp).await;
+    assert!(result1.is_ok());
+
+    let result2 = rc.list_global_runtime_parameters().await;
+    assert!(result2.is_ok());
+
+    let result3 = rc.clear_global_runtime_parameter("cluster_tags").await;
+    assert!(result3.is_ok());
+}
+
+#[tokio::test]
 async fn test_async_deserialize_sequence_value() {
     let json = r#"
       {
@@ -104,7 +219,7 @@ async fn test_async_deserialize_sequence_value() {
     assert_eq!(param.vhost, "test");
     assert_eq!(param.component, "limits");
 
-    let expected_value: RuntimeParameterValue = serde_json::Map::new();
+    let expected_value = serde_json::Map::new();
 
     assert_eq!(param.value.0, expected_value);
 }