+++ tests/async_virtual_host_tests.rs
@@ -70,6 +70,36 @@ async fn test_async_create_vhost() {
 }
 
 #[tokio::test]
+async fn test_async_create_vhost_without_dqt() {
+    let endpoint = endpoint();
+    let rc = Client::new(&endpoint, USERNAME, PASSWORD);
+    let name = "rust_test_async_create_vhost_without_dqt";
+
+    let _ = rc.delete_vhost(name, false).await;
+
+    let result1 = rc.get_vhost(name).await;
+    assert!(result1.is_err());
+
+    let desc = format!("{} description", &name);
+    let params = VirtualHostParams {
+        name,
+        description: Some(&desc),
+        tags: Some(vec!["tag-a", "tag-b"]),
+        default_queue_type: None,
+        tracing: false,
+    };
+    let result2 = rc.create_vhost(&params).await;
+    assert!(result2.is_ok());
+
+    let result3 = rc.get_vhost(name).await;
+    assert!(result3.is_ok());
+    let vh2 = result3.unwrap();
+    assert!(vh2.name == name);
+
+    let _ = rc.delete_vhost(name, false).await;
+}
+
+#[tokio::test]
 async fn test_async_update_vhost() {
     let endpoint = endpoint();
     let rc = Client::new(&endpoint, USERNAME, PASSWORD);